Ensuring identifier is always an integer
This fixes a bug I was getting when editing a file, where the identifier was a string. This caused a 'no implicit conversion of String to Integer' bug in the #write method when calling connection.lo_open. On a side note, the uploader was on an ActiveAdmin form (not sure if that had something to do with it).